added remove and add resources / added helper for check if enough resources are there
