+++ models/game/index.js
@@ -4,8 +4,8 @@ const { health } = require('./components/health')
 const { collector } = require('./components/collector')
 const { markerSpawner, towerSpawner, robotSpawner } = require('./components/spawner')
 const { shooter } = require('./components/shooter')
-const { resourceStore } = require('./components/resourceStore')
 const reducers = require('./reducers')
+const effects = require('./effects')
 const initialState = require('./initial-state')
 
 module.exports = game.model({
@@ -17,8 +17,8 @@ module.exports = game.model({
     markerSpawner,
     towerSpawner,
     robotSpawner,
-    shooter,
-    resourceStore
+    shooter
   },
+  effects,
   reducers
 })
